requirements.txt
# ================================================================
# 🧾 RTM Automation – Full Pipeline Dependencies
# ------------------------------------------------
# Supports:
#   1️⃣ fetch_rtm_data.py           (Jira RTM API integration)
#   2️⃣ generate_rtm_report.py      (HTML/PDF report generation)
#   3️⃣ confluence_publish.py       (Confluence REST API upload)
#   4️⃣ send_email.py               (SMTP email notification)
#   5️⃣ rtm_export_selenium.py      (optional – Selenium browser automation)
#
# Tested Environment:
#   Python >= 3.10
#   Jenkins Windows (java -jar jenkins.war)
#   Chrome/Chromedriver >= 130.x
# ================================================================

# --------------------------
# 🌐 HTTP / REST API Clients
# --------------------------
requests==2.32.3
certifi>=2024.7.4

# --------------------------
# 📊 Data & Report Formatting
# --------------------------
tabulate==0.9.0
fpdf2==2.8.1

# --------------------------
# 🧠 Automation & Web Drivers
# --------------------------
selenium==4.38.0   # Used by rtm_export_selenium.py for browser-based RTM export

# --------------------------
# ⚙️ Spreadsheet Handling (optional)
# --------------------------
pandas==2.2.3
openpyxl==3.1.5

# --------------------------
# 🧰 Utility Libraries
# --------------------------
typing_extensions>=4.10.0
urllib3>=2.0.7

# ================================================================
# Notes:
#   • Keep all dependencies pinned to ensure reproducible builds.
#   • Use “python -m venv .venv && .venv\Scripts\pip install -r requirements.txt”
#     when initializing Jenkins virtual environments.
# ================================================================
